refactor: reorganize type imports and create type definitions
