+++ packages/serverless-typespec-generator/src/ir/type.test.ts
