+++ packages/serverless-typespec-generator/src/ir/type.ts
